network-error-logging/sends-report-on-success.https.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test that NEL reports are sent for successful requests</title>
  <script src='/resources/testharness.js'></script>
  <script src='/resources/testharnessreport.js'></script>
  <script src='./support/fetch.sub.js'></script>
</head>
<body>
  <script>
    obtainNELLock();
    promise_test(async t => {
      await clearReportingAndNELConfigurations();
      // Make a request to a resource whose response headers include a NEL
      // policy.
      await fetchResourceWithBasicPolicy();
      // That policy should apply to the request that delivered it, so the
      // collector should have received a report about the request.
      assert_true(await reportExists({
        url: getURLForResourceWithBasicPolicy(),
        user_agent: navigator.userAgent,
        type: "network-error",
        body: {
          method: "GET",
          sampling_fraction: 1.0,
          status_code: 200,
          phase: "application",
          type: "ok",
        },
        metadata: {
          content_type: "application/reports+json",
        },
      }));
    });
    releaseNELLock();
  </script>
</body>
</html>
